[refactor] 肥大化したDashboardのリファクタリング
