+++ front/constants/cols.ts
@@ -1,0 +1,2 @@
+// front/constants/cols.ts
+export const COLS = { lg: 72, md: 60, sm: 36, xs: 24, xxs: 12 };
