+++ front/types/LayoutItem.ts
@@ -1,0 +1,11 @@
+// front/types/LayoutItem.ts
+// グリッドレイアウトのアイテム
+export type LayoutItem = {
+  i: string;
+  x: number;
+  y: number;
+  w: number;
+  h: number;
+  symbol: string;
+  label: string;
+};
